Clean Up and Organize the committed
The code was using the tab on 4, I created a prettierrc file to make it a rule, and a prettify the code
I change some names of vars, such as Persist to PersistProps and persist to PersistObjects
I improve some legibility, and a change the README, removing an error and separate to codes that were on the same sheet, to be on two sheets of code snippet.
So, I think it was a Clean Up, Prettify, Reorganize, and beautify the code, to follow the beauty package.